use eventsource-parser
