+++ apps/main/src/app/api/workflows/course-generation/status/route.ts
@@ -15,6 +15,6 @@ export async function GET(request: Request) {
   });
 
   return new Response(stream, {
-    headers: { "Content-Type": "text/plain; charset=utf-8" },
+    headers: { "Content-Type": "text/event-stream" },
   });
 }
